added todouble func to double datatypes in models

+++ songs_app/lib/models/songs.dart
@@ -89,7 +89,7 @@ class Song {
   Song.fromMaptoSong(Map<String, dynamic> map) {
     this._songId = map["songId"];
     this._title = map["title"];
-    this._length = map["length"];
+    this._length = map["length"].toDouble();
     this._location = map["location"];
     this._albumId = map["albumId"];
     this._genreId = map["genreId"];
